Fix various issues.
LPCM parser: WIP fix PES packets header.
HDMV parser: Fix dirty read.
H.264 parser: Fix of HRD-Verifier (remaining issues with DPB accuracy).

+++ src/codec/lpcm/lpcm_parser.h
@@ -149,8 +149,8 @@ typedef struct {
 } WaveFmtChunk;
 
 typedef struct {
-  size_t fileSize;
-  size_t dataSize;
+  uint32_t fileSize;
+  uint32_t dataSize;
   uint32_t pesPacketLength;
 
   WaveFmtChunk fmt;
